+++ app/core/erp/class_views/buy/views.py
@@ -48,10 +48,13 @@ class BuyListView(LoginRequiredMixin, ListView):
                 for i in DetBuy.objects.filter(buy_id=request.POST['id']): 
                     data.append(i.toJSON())  
             elif action == 'delete':
-                cli = Buy.objects.get(pk=request.POST['id'])
-                cli.user_update = request.user.username
-                cli.save()
-                cli.delete()
+                if request.user.is_superuser:
+                    cli = Buy.objects.get(pk=request.POST['id'])
+                    cli.user_update = request.user.username
+                    cli.save()
+                    cli.delete()
+                else:
+                    data['error'] = 'No tiene permiso para ingresar a este modulo'
             else:
                 data['error'] = 'Ha ocurrido un error'
         except Exception as e:
