+++ app/core/erp/class_views/production/views.py
@@ -44,10 +44,13 @@ class ProductionListView(LoginRequiredMixin, ListView):
                 for i in DetProduction.objects.filter(crea_id=request.POST['id']):  
                     data.append(i.toJSON())  
             elif action == 'delete':
-                cli = Production.objects.get(pk=request.POST['id'])
-                cli.user_update = request.user.username
-                cli.save()
-                cli.delete()
+                if request.user.is_superuser:
+                    cli = Production.objects.get(pk=request.POST['id'])
+                    cli.user_update = request.user.username
+                    cli.save()
+                    cli.delete()
+                else:
+                    data['error'] = 'No tiene permiso para ingresar a este modulo'
             else:
                 data['error'] = 'Ha ocurrido un error'
         except Exception as e:
