+++ app/core/erp/class_views/supplier/views.py
@@ -48,10 +48,13 @@ class SupplierListView(LoginRequiredMixin, ListView):
                 cli.user_update = request.user.username
                 cli.save()
             elif action == 'delete':
-                cli = Supplier.objects.get(pk=request.POST['id'])
-                cli.user_update = request.user.username
-                cli.save()
-                cli.delete()
+                if request.user.is_superuser:
+                    cli = Supplier.objects.get(pk=request.POST['id'])
+                    cli.user_update = request.user.username
+                    cli.save()
+                    cli.delete()
+                else:
+                    data['error'] = 'No tiene permiso para ingresar a este modulo'
             else:
                 data['error'] = 'Ha ocurrido un error'
         except Exception as e:
